+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 Unreleased
 
 - Made it possible to use uppercase extensions.
-- Added rule-file option in `draw` command.
+- Added rule-file option in `analyze` and `draw` commands.
 - Fixed bug without rule option in `analyze` command.
 
 ## 0.6.3
@@ -19,7 +19,7 @@ Released 2020-01-20
 Released 2020-01-20
 
 - Added alias command `amcli`.
-- Added rule option in `analyze` and `draw` command.
+- Added rule option in `analyze` and `draw` commands.
 
 ## 0.6.1